fix: GeoJSON layer with multiple markers  fails
A geojson layer with multiple Marker icons failed due to a bug in code. The bug was associated with a misplaced opening bracket. 
As a note: Leaflet's geoJSON expects the first param to be the geojson object and the second param to be the config string. This was incorrect in the multiple map markers branch of the SetupGeojsonLayer code.